Migration to firestore

+++ app/src/main/java/com/carista/ui/main/fragments/PostFragment.java
@@ -18,13 +18,19 @@ import com.carista.data.db.AppDatabase;
 import com.carista.data.realtimedb.models.PostModel;
 import com.carista.utils.Device;
 import com.google.android.material.snackbar.Snackbar;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
+import com.google.firebase.firestore.CollectionReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.EventListener;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.FirebaseFirestoreException;
+import com.google.firebase.firestore.Query;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
+import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.ArrayList;
+
+import static android.widget.NumberPicker.OnScrollListener.SCROLL_STATE_IDLE;
+import static com.carista.photoeditor.photoeditor.TextEditorDialogFragment.TAG;
 
 public class PostFragment extends Fragment {
 
@@ -32,6 +38,7 @@ public class PostFragment extends Fragment {
     private RecyclerView recyclerView;
     private String lastLazyItem;
     private String lastLazyKey;
+    private long oldtimestamp;
 
     public PostFragment() {
     }
@@ -57,12 +64,19 @@ public class PostFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        FirebaseDatabase database = FirebaseDatabase.getInstance();
-        DatabaseReference databaseReference = database.getReference("posts");
-        if (Device.isNetworkAvailable(getContext()))
-            databaseReference.orderByKey().limitToLast(5).addListenerForSingleValueEvent(new ValueEventListener() {
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+        CollectionReference query = db.collection("posts");
+
+        if (Device.isNetworkAvailable(getContext())) {
+
+            query.orderBy("timestamp", Query.Direction.DESCENDING).limit(3).addSnapshotListener(new EventListener<QuerySnapshot>() {
                 @Override
-                public void onDataChange(DataSnapshot dataSnapshot) {
+                public void onEvent(@Nullable QuerySnapshot value, @Nullable FirebaseFirestoreException error) {
+                    if(error != null){
+                        Log.w("LIST_POSTS", "listen:error", error);
+                        return;
+                    }
+                    adapter.clearData();
                     try {
                         Thread thread = new Thread(() -> AppDatabase.getInstance().postDao().deleteAll());
                         thread.start();
@@ -71,79 +85,63 @@ public class PostFragment extends Fragment {
                         e.printStackTrace();
                     }
 
-                    ArrayList<PostModel> postModels = new ArrayList<>();
-                    ArrayList<String> postKeys = new ArrayList<>();
-
-                    for (DataSnapshot post : dataSnapshot.getChildren()) {
-                        String id = post.getKey();
-                        PostModel postModel = new PostModel(id, post.getValue());
-                        postModels.add(postModel);
-                        postKeys.add(id);
+                    for(DocumentSnapshot doc : value.getDocuments()){
+                        String id = doc.getId();
+                        PostModel postModel = new PostModel(id, doc.getData());
+                        oldtimestamp = postModel.timestamp;
+                        adapter.addPost(postModel);
+                        AppDatabase.executeQuery(() -> AppDatabase.getInstance().postDao().insertAll(postModel));
                     }
 
-                    for (int i = postModels.size() - 1; i >= 0; i--) {
-                        adapter.addPost(postModels.get(i));
-                        final int j = i;
-                        AppDatabase.executeQuery(() -> AppDatabase.getInstance().postDao().insertAll(postModels.get(j)));
-                        lastLazyItem = postKeys.get(i);
-                    }
-                }
 
-                @Override
-                public void onCancelled(DatabaseError error) {
-                    // Failed to read value
-                    Log.w("ERROR", "Failed to read value.", error.toException());
                 }
             });
 
-        recyclerView.addOnScrollListener(new RecyclerView.OnScrollListener() {
+        }
+
+        recyclerView.setOnScrollListener(new RecyclerView.OnScrollListener() {
+            private int currentVisibleItemCount;
+            private int currentScrollState;
+            private int currentFirstVisibleItem;
+            private int totalItem;
+            @Override
+            public void onScrollStateChanged(@NonNull RecyclerView recyclerView, int newState) {
+                super.onScrollStateChanged(recyclerView, newState);
+                this.currentScrollState = newState;
+                this.isScrollCompleted();
+            }
+
+
             @Override
             public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
                 super.onScrolled(recyclerView, dx, dy);
-                if (!Device.isNetworkAvailable(getContext()))
-                    return;
-                if (!recyclerView.canScrollVertically(1) && recyclerView.canScrollVertically(-1)) {
-                    if (lastLazyItem != null) {
-                        FirebaseDatabase database = FirebaseDatabase.getInstance();
-                        DatabaseReference databaseReference = database.getReference("posts");
-                        databaseReference.orderByKey().endAt(lastLazyItem).limitToLast(6).addListenerForSingleValueEvent(new ValueEventListener() {
-                            @Override
-                            public void onDataChange(DataSnapshot dataSnapshot) {
-                                try {
-                                    Thread thread = new Thread(() -> AppDatabase.getInstance().postDao().deleteAll());
-                                    thread.start();
-                                    thread.join();
-                                } catch (InterruptedException e) {
-                                    e.printStackTrace();
-                                }
-
-                                ArrayList<PostModel> postModels = new ArrayList<>();
-                                ArrayList<String> postKeys = new ArrayList<>();
-
-                                for (DataSnapshot post : dataSnapshot.getChildren()) {
-                                    String id = post.getKey();
-                                    PostModel postModel = new PostModel(id, post.getValue());
-                                    postModels.add(postModel);
-                                    postKeys.add(id);
-                                }
-
-                                for (int i = postModels.size() - 2; i >= 0; i--) {
-                                    adapter.addPost(postModels.get(i));
-                                    final int j = i;
-                                    AppDatabase.executeQuery(() -> AppDatabase.getInstance().postDao().insertAll(postModels.get(j)));
-                                    lastLazyItem = postKeys.get(i);
-                                }
-                            }
-
-                            @Override
-                            public void onCancelled(DatabaseError error) {
-                                // Failed to read value
-                                Log.w("ERROR", "Failed to read value.", error.toException());
-                            }
-                        });
-                    }
-                }
             }
+
+
+            private void isScrollCompleted () {
+                if (totalItem - currentFirstVisibleItem == currentVisibleItemCount
+                        && this.currentScrollState == SCROLL_STATE_IDLE) {
+                    query.orderBy("timestamp", Query.Direction.DESCENDING).startAt(oldtimestamp).limit(3).addSnapshotListener(new EventListener<QuerySnapshot>() {
+                        @Override
+                        public void onEvent(@Nullable QuerySnapshot value, @Nullable FirebaseFirestoreException error) {
+                            if(error != null){
+                                Log.w("LIST_POSTS", "listen:error", error);
+                                return;
+                            }
+                            for(DocumentSnapshot doc : value.getDocuments()){
+                                String id = doc.getId();
+                                PostModel postModel = new PostModel(id, doc.getData());
+                                if(postModel.timestamp == oldtimestamp)
+                                    continue;
+                                oldtimestamp = postModel.timestamp;
+                                adapter.addPost(postModel);
+                                AppDatabase.executeQuery(() -> AppDatabase.getInstance().postDao().insertAll(postModel));
+                            }
+                        }
+                    });
+
+                }
+            };
         });
